pyproject.toml
[project]
name = "paper-search"
version = "0.1.3"
description = "AI-powered academic paper search workflow system"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx",
    "python-dotenv",
    "openai",
    "anthropic",
    "google-genai",
    "google-search-results",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]
mcp = [
    "mcp>=1.22",
]

[project.scripts]
paper-search-mcp = "paper_search.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_search"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
